refactor(office): split OfficeScene into FloorLayer + SpriteLayer
Phase 1 pure refactor — behaviour-preserving. OfficeScene.ts shrinks from
556 lines to 187 lines by extracting two new layer files:

  game/layers/types.ts      — shared payload interfaces (breaks circular dep)
  game/layers/FloorLayer.ts — floors, desks, click zones, idle indicators, stairs
  game/layers/SpriteLayer.ts — sprites, walks, indicator swaps, desk stacking

OfficeScene becomes a thin composition root: lifecycle, input, camera, and
delegating calls to the two layers. No behaviour change; all 28 unit tests
stay green; pnpm lint clean; no new dependencies.

Closes the Phase 1 acceptance criteria from:
docs/design/office-mode/vertical-slice-v1/phase-1-pure-refactor.md
