+++ apps/web/src/components/office/game/layers/types.ts
@@ -1,0 +1,31 @@
+// Shared payload types for the Office scene layers.
+//
+// Lives here (not in OfficeScene.ts) to break the potential circular-import
+// that would arise if FloorLayer / SpriteLayer imported from their parent scene
+// while OfficeScene imported from those layers. OfficeScene re-exports these
+// for backward-compat so the React side doesn't need import-path changes.
+
+export interface OfficeProject {
+  slug: string;
+  name: string;
+  color?: string;
+}
+
+export interface DeskClickPayload {
+  projectSlug: string;
+  workItemId: string;
+  externalId: string;
+  title?: string;
+}
+
+export interface FloorChangePayload {
+  floorIndex: number;
+  projectSlug: string;
+  projectName: string;
+}
+
+/** Verified PNG assets to queue in preload (probed by the React mount). */
+export interface VerifiedAsset {
+  key: string;
+  url: string;
+}
